+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Complete list of releases [here](https://github.com/Keplerjs/Kepler/releases)
 - new method K.Profile.setMobile()
 - map gps fix gps user
 - plugin geoinfo in panelEdit Name tips by locality
+- K.Map.addGeojson fix params
 
 ## v1.5.7
 - plugin admin new page admin users